fixing colors
