+++ example/src/examples/HomeScreen.tsx
@@ -1,0 +1,90 @@
+import React, { useContext } from 'react';
+import { Text, StatusBar, StyleSheet } from 'react-native';
+import { TouchableOpacity, ScrollView } from 'react-native-gesture-handler';
+import { SafeAreaView } from 'react-native-safe-area-context';
+// eslint-disable-next-line import/no-unresolved
+import { Button, ThemeContext, colors } from 'supergroww';
+
+import { COMPONENT_SCREENS, ANIMATED_SCREENS } from './navigation';
+
+const HomeScreen = ({ navigation, route }: any) => {
+  const currentTheme = useContext(ThemeContext);
+  const themeColors = colors[currentTheme];
+  const { setTheme, incVal } = route.params;
+  const changeThemeClick = () => {
+    incVal();
+    if (currentTheme === 'LIGHT') {
+      setTheme('DARK');
+      StatusBar.setBarStyle('light-content', true);
+    } else {
+      setTheme('LIGHT');
+      StatusBar.setBarStyle('default', true);
+    }
+  };
+
+  return (
+    <SafeAreaView style={{ flex: 1, backgroundColor: themeColors.white }}>
+      <ScrollView
+        showsVerticalScrollIndicator={false}
+        contentContainerStyle={{ flexGrow: 1 }}
+      >
+        <Text
+          style={{
+            fontSize: 20,
+            margin: 20,
+            fontWeight: 'bold',
+            color: themeColors.font_1,
+          }}
+        >
+          UI components
+        </Text>
+        {Object.keys(COMPONENT_SCREENS).map((item) => (
+          <TouchableOpacity
+            onPress={() => navigation.navigate(COMPONENT_SCREENS[item])}
+            style={styles.navButtonContainer}
+            key={item}
+          >
+            <Text style={{ color: themeColors.font_1 }}>
+              {COMPONENT_SCREENS[item]}
+            </Text>
+          </TouchableOpacity>
+        ))}
+        <Text
+          style={{
+            fontSize: 20,
+            margin: 20,
+            fontWeight: 'bold',
+            color: themeColors.font_1,
+          }}
+        >
+          Animated components
+        </Text>
+        {Object.keys(ANIMATED_SCREENS).map((item) => (
+          <TouchableOpacity
+            onPress={() => navigation.navigate(ANIMATED_SCREENS[item])}
+            style={styles.navButtonContainer}
+            key={item}
+          >
+            <Text style={{ color: themeColors.font_1 }}>
+              {ANIMATED_SCREENS[item]}
+            </Text>
+          </TouchableOpacity>
+        ))}
+        <Button
+          onPress={changeThemeClick}
+          label="CHANGE THEME"
+          customStyles={{ margin: 20 }}
+        />
+      </ScrollView>
+    </SafeAreaView>
+  );
+};
+
+const styles = StyleSheet.create({
+  navButtonContainer: {
+    padding: 20,
+    borderBottomWidth: 1,
+    borderBottomColor: 'rgba(191, 191, 191, 0.5)',
+  },
+});
+export default HomeScreen;
